+++ tests/testapp/test_feincms3.py
@@ -630,7 +630,11 @@ class Test(TestCase):
             )),
             '/admin/',
         )
-        with self.assertRaises(NoReverseMatch):
+        with self.assertRaisesRegex(
+                NoReverseMatch,
+                "Reverse for any of 'not-exists-1', 'not-exists-2' with"
+                " arguments '\[\]' and keyword arguments '{}' not found."
+        ):
             reverse_any(('not-exists-1', 'not-exists-2'))
 
     def test_move_clean_and_save(self):